+++ app/Services/PaymentDetailsProcessor.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright (c) 2018-2022 Adshares sp. z o.o.
+ * Copyright (c) 2018-2023 Adshares sp. z o.o.
  *
  * This file is part of AdServer
  *
@@ -56,6 +56,7 @@ class PaymentDetailsProcessor
         $exchangeRate = $this->fetchExchangeRate();
         $feeCalculator = new PaymentDetailsFeeCalculator($this->fetchLicenseFee(), config('app.payment_rx_fee'));
         $totalLicenseFee = 0;
+        $totalOperatorFee = 0;
         $totalEventValue = 0;
 
         $exchangeRateValue = $exchangeRate->getValue();
@@ -86,10 +87,11 @@ class PaymentDetailsProcessor
             $case->networkCasePayments()->save($networkCasePayment);
 
             $totalLicenseFee += $calculatedFees['license_fee'];
+            $totalOperatorFee += $calculatedFees['operator_fee'];
             $totalEventValue += $eventValue;
         }
 
-        return new PaymentProcessingResult($totalEventValue, $totalLicenseFee);
+        return new PaymentProcessingResult($totalEventValue, $totalLicenseFee, $totalOperatorFee);
     }
 
     public function addAdIncomeToUserLedger(AdsPayment $adsPayment): void